% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPois-package.R
\docType{package}
\name{easyPois-package}
\alias{easyPois}
\alias{easyPois-package}
\title{easyPois}
\description{
The package calculates the maximum likelihood estimator with observed data that we assume to be distributed under Poisson distribution.
}
\examples{
y <- 1L:20L
estimatePois(y, "bootstrap", B = 1000, lambda = 2)



}
\author{
Cecilia Y. Sui: \email{c.sui@wustl.edu}
}
