% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPois-package.R
\docType{package}
\name{easyPois}
\alias{easyPois}
\alias{easyPois-package}
\title{easyPois}
\description{
The easyPois package is used to estimate the value of lambda from a vector of
data. You can create an S4 Class with the necessary information for a Poisson
Distribution.
}
\examples{
set.seed(12345)
y <- rpois(1000, 5)

est <- estimatePois(y, SEType = "basic")
estBoot <- estimatePois(y, SEType = "bootstrap", samples = 20)

}
\author{
Peter Bachman: \email{bachman.p@wustl.edu}
}
