% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\title{Estimate for a Poisson Distribution}
\usage{
estimatePois(y, SEType = "basic", samples = 10)
}
\arguments{
\item{y}{A vector with observed data}

\item{SEType}{Defines how the Standard Errors will be calculated. Can pick
from \code{basic} or \code{bootstrap}. Defaults to \code{bootstrap}.}

\item{samples}{If \code{SEType = "bootstrap"} then this defines the number of
samples that will be used.}
}
\value{
an S4 class of \code{PoisMLE}
}
\description{
Calculates the estimates for a Poisson Distribution
}
\seealso{
\link{\code{PoisMLE}}
}
\author{
Peter Bachman: \email{bachman.p@wustl.edu}
}
