% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Maximum Likelihood Estimator}
\usage{
mle(y)
}
\arguments{
\item{y}{A vector with the observed data}
}
\value{
The MLE Estimate for lambda
}
\description{
Calculates the MLE for a Poisson Distribution
}
\examples{
set.seed(12345)
y <- rpois(1000, 5)

mle(y)

}
\author{
Peter Bachman: \email{bachman.p@wustl.edu}
}
