% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPois.R
\docType{methods}
\name{plotPois}
\alias{plotPois}
\alias{plotPois,PoisMLE-method}
\title{Plot the MLE}
\usage{
plotPois(object)
}
\arguments{
\item{object}{A \code{PoisMLE} class object}
}
\value{
Plot of MLE with 95\\% confidence intervals
}
\description{
Plots the MLE with confidence intervals for \code{PoisMLE} class objects
}
\examples{
set.seed(12345)
y <- rpois(1000, 5)

est <- estimatePois(y, SEType = "basic")
estBoot <- estimatePois(y, SEType = "bootstrap", samples = 20)

plotPois(est)
plotPois(estBoot)

}
\seealso{
\linkS4class{PoisMLE}
}
\author{
Peter Bachman: \email{bachman.p@wustl.edu}
}
