% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\title{Poisson S4 Class}
\description{
Poisson S4 Class
}
\section{Slots}{

\describe{
\item{\code{y}}{vector of values}

\item{\code{MLE}}{maximum likelihood estimator}

\item{\code{LL}}{Log Likelihood from \code{y}}

\item{\code{SE}}{Standard Error}

\item{\code{SEType}}{Method of calculating Standard Error}
}}

