% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE_Class.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\title{PoisMLE}
\description{
This is the class for a poisson model
}
\details{
Fitted model objects have the following slots:
\itemize{
\item \code{y} The inputed data by a user
\item \code{MLE} the mle estimator calculated
\item \code{LL} the log likelihood calculated from the data
\item \code{SE} this is the standard error for the MLE
\item \code{SEtype} the method used to calculate the se
}
}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{mle}}

\code{\link{standardError}}
}
\author{
Amaan Charaniya
}
