% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPois-package.R
\docType{package}
\name{easyPois}
\alias{easyPois}
\alias{easyPois-package}
\title{easyPois}
\description{
The easyPois package is used to estimate the value of lambda from a vector of
data. You can create an S4 Class with the necessary information for a Poisson
Distribution.
}
\author{
Peter Bachman: \email{bachman.p@wustl.edu}
}
