% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\title{Fit \code{PoisMLE} model}
\usage{
estimatePois(y, SEtype, B = 10)
}
\arguments{
\item{y}{A vector of positive integer observations.}

\item{SEtype}{A character string of \code{"basic"} or \code{"bootstrap"}; see \code{standardError}.}

\item{B}{An integer; the default is set to \code{B=10}. This argument is not used
if \code{SEtype="basic"}.}
}
\value{
An object of class \code{'PoisMLE'} containing
\item{y}{The original vector of positive integer observations.}
\item{MLE}{The maximum likelihood estimator for given observations.}
\item{LL}{The log likelihood calculated from the observed data using the MLE estimator.}
\item{SE}{The standard error for the MLE.}
\item{SEtype}{The method used to calculate \code{SE}, supplied by the user.}
}
\description{
For a vector of observed data, returns an object of class \code{PoisMLE},
including an estimate of its associated \eqn{\lambda} parameter.
}
\examples{
\donttest{y_vec <- c(1,2,3,4)
estimatePois(y_vec, "bootstrap", B=10)}

}
\seealso{
PoisMLE-class, mle, logLik
}
\author{
Alma Velazquez
}
