% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\alias{estimatePois-method}
\title{Estimating Poisson Distribution}
\usage{
estimatePois(y, SEtype, B, lambda)
}
\arguments{
\item{y}{The vector of observed data.}

\item{SEtype}{The method used to calculate the standard error. It can take on two values: basic and bootstrap.}

\item{B}{The number of bootstrapped resamplings.}

\item{lambda}{The assumed value of \eqn{\lambda}.}
}
\value{
An object of class PoisMLE containing: 
\item \code{y} {The original data}
\item \code{MLE} {The maximum likelihood estimator for this dataset}
\item \code{LL} {The log likelihood calculated from the observed data assuming the MLE is correct}
\item \code{SE} {The standard error for the MLE}
\item \code{SEtype} {The method used to calculate the standard error}
}
\description{
Estimates Poisson Distribution
}
\examples{
estimatePois(y, SEtype = "basic", B = 500, lambda = 2)

}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{mle}}, \code{\link[PoisMLE]{standardError}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Cecilia Y. Sui: \email{c.sui@wustl.edu}
}
