% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\title{estimatePois function for MLE calculation}
\usage{
estimatePois(y, lambda, SEtype = c("basic", "bootstrap"), B = 200)
}
\arguments{
\item{y}{Observed data in vector form}

\item{lambda}{Designated value of lambda for the Poisson distribution}

\item{SEtype}{Method of calculating the standard error, either \code{basic} or \code{bootstrap}.}

\item{B}{Number of resamplings to be done if bootrapping the standard errors.Default is set to 200.}
}
\value{
An object of class PoisMLE
\item{y}{The original data}
\item{MLE}{The maximum likelihood estimator for this dataset}
\item{LL}{The log likelihood calculated from the observed data assuming the MLE is correct}
\item{SE}{The standard error for the MLE}
\item{SEtype}{The method used to calculate the standard error}
}
\description{
Function for calculating log likelihood, MLE, and standard errors of a Poisson distribution data
}
\examples{
set.seed(10)
y <- sample(x=1:10, size=10, replace=TRUE)
estimatePois(y, lambda = 10, SEtype = "bootstrap")
estimatePois(y, lambda = 10, SEtype= "basic")
}
\author{
Eunsan Jo
}
