% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\title{estimatePois}
\usage{
estimatePois(y, type)
}
\arguments{
\item{y}{this is a numeric vector of your observed data. That's all you have to give and this will do the rest.}

\item{type}{this is the type of se you want, either "basic" or "bootstrapped"}
}
\value{
I'll provide you a few things in a single object:
 \item{y}{you'll get back your data in case your forgot about it}
 \item{mle}{you'll get the mle of your data}
 \item{ll}{you'll get the loglikelihood of your data}
 \item{se}{this is the standard error for the MLE}
 \item{se}{the method used to calculate the se}
}
\description{
this fuction will take in your data and do all the work for you if you're too lazy to use the individual functions in this package
}
\note{
this function will return an S4 class of PoisMLE
}
\examples{

y <- rpois(100,50)
estimatePois(y, "basic")
estimatePois(y, "bootstrap")


}
\seealso{
\code{\link{standarderror}}

\code{\link{mle}}

\code{\link{logLik}}
}
\author{
Amaan Charaniya
}
