% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\alias{estimatePois,ANY-method}
\title{Estimating the Poisson Distribution}
\usage{
estimatePois(y, SEtype, B)
}
\arguments{
\item{y}{A vector of values}

\item{SEtype}{The method used to calculate the standard error: basic or bootstrap}

\item{B}{The number of bootstrapped resamplings}
}
\value{
Returns an object of class \code{'PoisMLE'}, which contains:
\item{y}{The original vector of data}
\item{MLE}{The maximum likelihood estimator for the data}
\item{LL}{The log likelihood calculated from the observed data}
\item{SE}{The standard error for the MLS}
\item{SEtype}{The method used to calculate the standard error: basic or bootstrap}
}
\description{
Calculates the log likelihood, the maximum likelihood, and the standard error for a vector of values.
}
\seealso{
PoisMLE-class
}
\author{
Kimberly Acquilano

#' @examples

estimatePois(1:10, "bootstrap", 10)
estimatePois(1:10, "basic", 10)
}
