% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\title{Estimate for a Poisson Distribution}
\usage{
estimatePois(y, SEtype = "bootstrap", B = 100)
}
\arguments{
\item{y}{Original data; vector of non-negative integers.}

\item{SEtype}{Tells \code{standardError} function to use the \code{basic} or \code{bootstrap} method for calculating the standard error. \code{standardError} defaults to \code{bootstrap}.}

\item{B}{Number of samples for \code{bootstrap} method. The default is 100 samples.}
}
\value{
Object of class \code{PoisMLE}
}
\description{
Function that estimates and returns relevant statistics for a Poisson
distribution from observed data.
}
\examples{
y <- rpois(2000,3)
estPois <- estimatePois(y = y, SEtype = "basic")
estPoisBootstrap <- estimatePois(y = y, SEtype = "bootstrap", B = 200)

}
\author{
Patrick Edwards
}
