% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\alias{estimatePois,ANY-method}
\title{Estimate the Poisson model}
\usage{
estimatePois(y, SEtype, B = 20)

\S4method{estimatePois}{ANY}(y, SEtype, B = 20)
}
\arguments{
\item{y}{is a vector for which all entries are strictly non-negative reals}

\item{SEtype}{Standard error calculation method}

\item{B}{Is an integer that I have set to 20 as the default. This is not used when \code{"basic"} is selected.}
}
\value{
Object in \code{"PoisMLE"} with the following:
#'\itemize{
\item \code{y} is a vector for which all entries are strictly non-negative reals
\item \code{MLE} Poisson MLE for an entry of y
\item \code{LL} Log likelihood from data in y using Poisson MLE estimator
\item \code{SE} Standard error of Poisson MLE
\item \code{SEtype} Standard error calculation method
}
}
\description{
For a vector of strictly non-negative entries, this returns an object in the class of
\code{PoisMLE} and \eqn{\hat \lambda}.
}
\examples{
\donttest{the_y<-c(1:20)
estimatePois(the_y, SEtype="basic", B=20)}

}
\author{
Zion Little
}
