% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Calculate a log likelihood in \code{PoisMLE} object}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{A vector of positive integer observations.}

\item{lambda}{The assumed value of \eqn{\lambda}, see \code{mle}.}
}
\value{
A log likelihood for consumption of objects of class \code{PoisMLE}
\item{LL}{The log likelihood value}
}
\description{
For internal use only. This function calculates the log likelihood for a given
vector of observed data, conditional on the value of the \eqn{\lambda} parameter
in the associated Poisson distribution. The calculation is performed according to
\deqn{LL(\lambda) = -n\lambda-\sum^n_[i=1]\ln(y_i!) + \ln(\lambda)\summ^n_[i=1]y_i}.
The computed value is ultimately returned when initializing an object of class \code{PoisMLE}.
}
\examples{
\donttest{y_vec <- c(1,2,3,4)
logLik(y_vec, mle(y_vec))}
}
\seealso{
mle, PoisMLE-class
}
\author{
Alma Velazquez
}
\keyword{internal}
