% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Log Likelihood}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{a vector of positive values}

\item{lambda}{value for the expected lambda}
}
\value{
log likelihood value
}
\description{
Calculating the Log Likelihood for a Poisson Model
}
