% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLik-method}
\title{Calculating log likelihood}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{The vector of observed data.}

\item{lambda}{The assumed value of \eqn{\lambda}.}
}
\value{
The function outputs the log likelihood for the observed data conditioned on the value of \eqn{\lambda}.
}
\description{
Calculates a log likelihood for the observed data based on Poisson distribution.
}
\examples{
y <- c(1L:10L)
logLik(y, lambda = 2)

}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{mle}}, \code{\link[PoisMLE]{standardError}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Cecilia Y. Sui: \email{c.sui@wustl.edu}
}
