% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{loglik}
\title{Log Likelihood of a Poisson Distribution}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{The vector of observed data}

\item{lambda}{The assumed value of lambda}
}
\value{
the log likelihood for the observed dataset y conditioned on the value of lambda
}
\description{
Calculates a log likelihood for the observed data
}
\note{
part of easyPois package
}
\examples{
set.seed(10)
y <- sample(x=1:10, size=2000, replace=TRUE)
logLik(y, 10)
}
\seealso{
\code{mle}, \code{se}
}
\author{
Eunsan Jo <\email{ejo@wustl.edu}>
}
