% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{logLik}
\usage{
logLik(y, L)
}
\arguments{
\item{y}{this is a numeric vector of your observed data, please don't put characters in here.}

\item{L}{this is your lambda. You need this. you can calculate it with the mle function in this package but you need a lambda}
}
\value{
A single numeric output:
 \item{mle}{you'll get the loglikelihood of your data if you did this right}
}
\description{
this fuction will generate the log likelihood for your data
}
\note{
make sure you have a lambda
}
\examples{

y <- rpois(100,50)
L <- mle(y)
logLik(y, L)

}
\seealso{
\code{\link{estimatePois}}

\code{\link{mle}}

\code{\link{standardError}}
}
\author{
Amaan Charaniya
}
