% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Calculate logged likelihood}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{Original data; vector of non-negative integers}

\item{lambda}{The assumed value of the lambda estimate}
}
\value{
An object \code{LL} containing the log likelihood estimate
}
\description{
Calculates logged likelihood of the observed data given a value of lambda
}
\examples{

ytest <- rpois(2000,3)
logLik(y = ytest, lambda = 5)

}
\author{
Patrick Edwards
}
