% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Log Likelihood estimator}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{A vector with the observed data}

\item{lambda}{A value for the expected lambda}
}
\value{
The Log Likelihood Estimate
}
\description{
Calculates the Log Likelihood from observed data and a value for lambda
}
\examples{
set.seed(12345)
y <- rpois(1000, 5)
l <- mle(y)

logLik(y, l)

}
\seealso{
\code{\link[=mle]{mle()}}
}
\author{
Peter Bachman: \email{bachman.p@wustl.edu}
}
