% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Log Likelihood for Poisson Model}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{is a vector for which all entries are strictly non-negative reals}

\item{lambda}{is the assumed rate for a Poisson process or Poisson distribution}
}
\value{
The log likelihood value for objects in the class \code{PoisMLE}
}
\description{
This calculates the log likelihood in a \code{PoisMLE} object. This is for internal use.
The following function calculates log likelihood for a vector of given data where rate parameter
\eqn{\lambda} is assumed. This function performs the log likelihood equation given by
\deqn{LL(\lambda)=-n\lambda-\sum^n_[i=1]\ln(y_i!)+\ln(y_i!)+\ln(\lambda)\sum^n_[i=1]y_i}.
The value that the below function gives is returned when an object in the class \code{PoisMLE} is initialized.
}
\seealso{
PoisMLE-class, MLE,StandardError
}
\author{
Zion Little
}
\keyword{internal}
