% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Calculate MLE estimator for \eqn{\lambda} in \code{PoisMLE} object}
\usage{
mle(y)
}
\arguments{
\item{y}{A vector of positive integer observations.}
}
\value{
A maximum likelihood estimator for consumption of objects of class \code{PoisMLE}
\item{MLE}{The maximum likelihood estimator for \eqn{\lambda}}
}
\description{
For internal use only. This function calculates the maximum likelihood estimator for
the \eqn{\lambda} parameter in a Poisson distribution, such that
\deqn{\Pr(Y_i =y) = \frac{\lambda^{y_i} exp^{-\lambda}}{y_i!}}.
The estimator is then used and ultimately returned when initializing an object of class \code{PoisMLE}.
}
\examples{
\donttest{y_vec <- c(1,2,3,4)
mle(y_vec)}
}
\seealso{
standardError, logLik, PoisMLE-class
}
\author{
Alma Velazquez
}
\keyword{internal}
