% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Maximum Likelihood Estimate}
\usage{
mle(y)
}
\arguments{
\item{y}{a vector of positive values}
}
\value{
maximum likelihood estimate
}
\description{
Calculates the maximum likelihood estimate for a Possion Model
}
