% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\alias{mle-method}
\title{Calculating maximum likelihood estimator}
\usage{
mle(y)
}
\arguments{
\item{y}{The vector of observed data.}
}
\value{
The function outputs the maximum likelihood estimator for \eqn{\lambda}.
}
\description{
Calculates the maximum likelihood estimator for lambda.
}
\examples{
y <- 1L:20L
mle(y)

}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{mle}}, \code{\link[PoisMLE]{standardError}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Cecilia Y. Sui: \email{c.sui@wustl.edu}
}
