% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\alias{poismle}
\title{Returns Maximum Likelihood of a Poisson Distribution}
\usage{
mle(y)
}
\arguments{
\item{y}{Given observed data}
}
\value{
maximum likelihood estimator for the observed dataset y
}
\description{
Calculates the maximum likelihood estimator for the given dataset
}
\examples{
y = c(1:10)
mle(y)
}
\seealso{
\code{logLik}, \code{standardError}
}
\author{
Eunsan Jo <\email{ejo@wustl.edu}>
}
