% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{mle}
\usage{
mle(y)
}
\arguments{
\item{y}{this is a numeric vector of your observed data, please don't put characters in here. That's all you need}
}
\value{
A single numeric output:
 \item{mle}{you'll get the mle of your data if you did this right}
}
\description{
this fuction will generate the mle for your data (this can be your lambda)
}
\note{
This can be used as your lambda
}
\examples{

y <- rpois(100,50)
mle(y)

}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{standardError}}
}
\author{
Amaan Charaniya
}
