% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Calculate the maximum likelihood estimator}
\usage{
mle(y)
}
\arguments{
\item{y}{Original data; vector of non-negative integers.}
}
\value{
Object \code{MLE} containing the calculated maximum likelihood estimator.
}
\description{
Calculates the maximum likelihood estimator for lambda given observed data.
}
\examples{

ytest <- c(1:4,3,7:5,11)
mle(y = ytest)

}
\author{
Patrick Edwards
}
