% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{mle}
\alias{mle}
\alias{MLE}
\title{MLE for \equn{\lambda} in object of class \code{PoisMLE}}
\usage{
mle(y)
}
\arguments{
\item{y}{is a vector for which all entries are strictly non-negative reals}
}
\value{
MLE
}
\description{
For internal use. The function below runs the MLE calculation for Poisson rate \eqn{\lambda}. The following
calculation is performed:
\deqn{\frac{\lambda^{y_{i} e(-\lambda)} }{y_{i}!}}, which defines \deqn{\P(Y_{i}=y)}.
The estimator is used and returned when an object in the class \code{PoisMLE} is initialized.
}
\seealso{
PoisMLE-class, logLik,StandardError
}
\author{
Zion Little
}
\keyword{internal}
