% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLEClass.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{y}}{numeric.}

\item{\code{MLE}}{numeric.}

\item{\code{LL}}{numeric.}

\item{\code{SE}}{numeric.}

\item{\code{SEtype}}{character.}
}}

