% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLEClass.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\title{Poisson MLE Class}
\value{

}
\description{
An S4 class for a Possion Model
}
\section{Slots}{

\describe{
\item{\code{y}}{a vector of positive values}

\item{\code{MLE}}{The maximum likelihood estimator for the y vector.}

\item{\code{LL}}{The log likelihood calculated from the y vector.}

\item{\code{SE}}{The standard error for the MLE}

\item{\code{SEtype}}{The method used to calculate the standard error (either basic or bootstrap).}
}}

