% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE-class.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\alias{estimatePois-method}
\alias{initialize,PoisMLE-method}
\alias{plot,PoisMLE-method}
\title{A fitted MLE Poisson model}
\usage{
\S4method{initialize}{PoisMLE}(.Object, ...)

\S4method{plot}{PoisMLE}(x)
}
\description{
Objects of class \code{PoisMLE} are created by the \code{estimatePois} function
}
\details{
An object of the class `PoisMLE' has the following slots:
\itemize{
\item \code{y } The original vector of non-negative integer observations.
\item \code{MLE } The maximum likelihood estimator for given observations.
\item \code{LL } The log likelihood calculated from the observed data using the MLE estimator.
\item \code{SE } The standard error for the MLE.
\item \code{SEtype } The method used to calculate the standard error.
}
}
\author{
Alma Velazquez: \email{a.m.velazquez@wustl.edu}
}
