% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE.R
\name{check_PoisMLE}
\alias{check_PoisMLE}
\alias{PoisMLE-class}
\title{Define PoisMLE Class}
\usage{
check_PoisMLE(object)
}
\description{
An object of class \code{PoisMLE} can be created by the \code{estimatePois()} function
}
\details{
An object of class `PoisMLE` has the following slots:
(All descriptions are based on the exam project description file.)
\itemize{
\item \code{y} The original data
\item \code{MLE} The maximum likelihood estimator for this dataset
\item \code{LL} The log likelihood calculated from the observed data assuming the MLE is correct
\item \code{SE} The standard error for the MLE
\item \code{SEtype} The method used to calculate the standard error 
}
}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{mle}}, \code{\link[PoisMLE]{standardError}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Cecilia Y. Sui: \email{c.sui@wustl.edu}
}
