% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE.R
\docType{class}
\name{PoisMLE}
\alias{PoisMLE}
\title{Create Poisson MLE S4 Class Generator Function}
\description{
Object of class \code{PoisMLE} as created by the \code{PoisMLE} function. An object of class \code{PoisMLE} has the following slots:
}
\section{Slots}{

\describe{
\item{\code{y}}{Original data; vector of non-negative integers.}

\item{\code{MLE}}{The maximum likelihood estimator.}

\item{\code{LL}}{The log likelihood that was calculated from the observed data (assuming that the MLE is correct).}

\item{\code{SE}}{The MLE's standard error.}

\item{\code{SEtype}}{The method by which the standard errors are  This can be either \code{basic} or \code{bootstrap}.}
}}

\author{
Patrick Edwards
}
