% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE-class.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\alias{estimatePois-class}
\alias{initialize,PoisMLE-method}
\title{MLE for Poisson}
\usage{
\S4method{initialize}{PoisMLE}(.Object, ...)
}
\description{
Objects in class \code{PoisMLE} are created by the function \code{estimatePois}
}
\details{
Objects of class \code{PoisMLE} have the slots"
\itemize{
\item \code{y} is a vector for which all entries are strictly non-negative reals
\item \code{MLE} Poisson MLE for an entry of y
\item \code{LL} Log likelihood from data in y using Poisson MLE estimator
\item \code{SE} Standard error of Poisson MLE
\item \code{SEtype} Standard error calculation method
}
}
\author{
Zion Little: \email{l.zion@wustl.edu}
}
