% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\title{Calculate the standard error in \code{PoisMLE} object}
\usage{
standardError(y, SEtype, B = 10)
}
\arguments{
\item{y}{A vector of positive integer observations.}

\item{SEtype}{A character string of \code{"basic"} or \code{"bootstrap"}; see 'Details'.}

\item{B}{An integer; the default is set to \code{B=10}. This argument is not used
if \code{SEtype="basic"}.}
}
\value{
The standard error of the MLE object in class \code{PoisMLE}
\item{SE}{The standard error}
}
\description{
For internal use only. This function calculates the standard error for the
maximum likelihood estimator in the \code{PoisMLE} class using one of two methods
as given by the user. See 'Arguments'. The computed value is ultimately returned when
initializing an object of class \code{PoisMLE}.
}
\details{
The user specifies which method to use for this calculation when initializing a \code{PoisMLE} object with the
\code{estimatePois} function. If \code{"basic"}, the calculation is performed according to
\deqn{\sqrt(\frac{MLE}{n})}. If \code{"bootstrap"}, the SE is the standard deviation of \code{B} bootstrapped
resamplings of \code{y}.
}
\examples{
\donttest{y_vec <- c(1,2,3,4)
standardError(y_vec, "bootstrap", B=10)}
}
\seealso{
mle, PoisMLE-class
}
\author{
Alma Velazquez
}
\keyword{internal}
