% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\alias{standardError-method}
\title{Calculating the standard error}
\usage{
standardError(y, SEtype, B)
}
\arguments{
\item{y}{The vector of observed data.}

\item{SEtype}{The method used to calculate the standard error. It can take on two values: basic and bootstrap.}

\item{B}{The number of bootstrapped re-samplings.}
}
\value{
The function outputs the standard error calculated via the chosen method.
}
\description{
Calculates the standard error based on a chosen method by the user.
}
\examples{
y <- sample(1L:20L, 100, replace = TRUE)
standardError(y, "basic", B = 500) 

}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{mle}}, \code{\link[PoisMLE]{standardError}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Cecilia Y. Sui: \email{c.sui@wustl.edu}
}
