% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standarderror.R
\name{standarderror}
\alias{standarderror}
\title{standarderror}
\usage{
standarderror(y, type)
}
\arguments{
\item{y}{this is a numeric vector of your observed data, please don't put characters in here.}

\item{type}{this is the type of standard error your'd like, basic or bootstrapped}
}
\value{
A single numeric output:
 \item{mle}{you'll get the standard error of your data}
}
\description{
this fuction will generate the standard error for your mle estimate
}
\note{
this function will produce a bootstrap standard error with 1000 samples by default. You can not change this because I do not trust you.
}
\examples{

y <- rpois(100,50)
standarderror(y, "basic")
standarderror(y, "bootstrap")

}
\seealso{
\code{\link{estimatePois}}

\code{\link{mle}}

\code{\link{logLik}}
}
\author{
Amaan Charaniya
}
