% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\title{Standard Error (Poisson Distribution)}
\usage{
standardError(y, SEtype, B = 100)
}
\arguments{
\item{y}{Original data; vector of non-negative integers.}

\item{SEtype}{Tells \code{standardError} function to use the \code{basic}
or \code{bootstrap} method for calculating the standard error.
\code{standardError} defaults to \code{bootstrap}.}

\item{B}{Number of samples for \code{bootstrap} method. The default is
100 samples.}
}
\value{
The standard error of the observed data.
}
\description{
This function calculates the standard error for observed data following
the Poisson distribution.
}
\examples{

testy <- rpois(2000,3)

standardError(y = testy, SEtype = "basic", B = 50)
standardError(y = testy, SEtype = "bootstrap", B = 500)

}
\author{
Patrick Edwards
}
