% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\title{Standard Error for Poisson Distribution}
\usage{
standardError(y, SEType = "basic", B = 10)
}
\arguments{
\item{y}{A vector with the observed data}

\item{SEType}{Calculate the standard error using either \code{basic} or \code{bootstrap}. Defaults to \code{basic}.}

\item{B}{Number of samples. Default is 10 samples}
}
\value{
The Standard Error for the observed data.
}
\description{
Calculates the Standard Error for a Poisson Distribution
}
\examples{
set.seed(12345)
y <- rpois(1000, 5)

standardError(y, "basic")
standardError(y, "bootstrap", 20)

}
\author{
Peter Bachman: \email{bachman.p@wustl.edu}
}
