% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardError.R
\name{standardError}
\alias{standardError}
\alias{StandardError}
\title{Calculate the standard error of \code{PoisMLE}}
\usage{
standardError(y, SEtype, B = 20)
}
\arguments{
\item{y}{is a vector for which all entries are strictly non-negative reals}

\item{SEtype}{is entries of either \code{"basic"} or \code{"bootstrap"}. See description.}

\item{B}{Is an integer that I have set to 20 as the default. This is not used when \code{"basic"} is selected.}
}
\value{
the standard errors
}
\description{
For internal use. The function below finds the standard error of the MLE for object of the class \code{PoisMLE}.
The user selects an argument for which method to use in the calculation. The two methods are \code{"basic"} and \code{\"bootstrap}.
The following calculations are run, respectively:
\deqn{\sqrt(\frac{MLE}{n})} ; standard deviation of \code{B} bootstraps of \code{y}.
The output value is returned when object in class \code{PoisMLE} is initialized.
}
\seealso{
PoisMLE-class, logLik, MLE
}
\author{
Zion Little
}
\keyword{internal}
